[expoview] Extract InstallationId fetch, migration and creation to a separate helper class
